+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+-   `conditionallyExecutePredicates` to `FormSubmissionEvent`
+
 ## 1.8.0 (2020-07-09)
 
 ### Added